Only send data for a specific UUID

+++ Dao/MainDao.cs
@@ -23,24 +23,24 @@ namespace ftDB.Dao
     {
         private readonly PostgressDBContext _context = context;
 
-        public async Task<List<ModelExercise>> GetExerciseListAsync(string searchInput)
+        public async Task<List<ModelExercise>> GetExerciseListAsync(string searchInput, string uuid)
         {
             if (searchInput != "")
             {
-                return await GetExerciseListWhenStringExistsAsync(searchInput);
+                return await GetExerciseListWhenStringExistsAsync(searchInput, uuid);
             }
 
-            return await GetExerciseListEmptyStringAsync();
+            return await GetExerciseListEmptyStringAsync(uuid);
 
         }
 
-        public async Task PostWorkoutAsync(RequestModelPostWorkout completedWorkout)
+        public async Task PostWorkoutAsync(RequestModelPostWorkout completedWorkout, string uuid)
         {
             using var transaction = _context.Database.BeginTransaction();
 
             try
             {
-                CompletedWorkout workoutToPost = new(completedWorkout.Date, completedWorkout.StartTime, completedWorkout.EndTime, completedWorkout.Duration, completedWorkout.Name); // Create new CompletedWorkout entity
+                CompletedWorkout workoutToPost = new(completedWorkout.Date, completedWorkout.StartTime, completedWorkout.EndTime, completedWorkout.Duration, completedWorkout.Name, uuid); // Create new CompletedWorkout entity
 
                 int workoutId = await PostCompletedWorkoutAsync(workoutToPost);
 
@@ -78,9 +78,10 @@ namespace ftDB.Dao
             }
         }
 
-        public async Task<List<ModelPastWorkout>> GetAllWorkoutsAsync()
+        public async Task<List<ModelPastWorkout>> GetAllWorkoutsAsync(string uuid)
         {
             List<ModelPastWorkout> workouts = await _context.CompletedWorkouts
+                .Where(w => w.Uuid == uuid)
                 .Include(cw => cw.ExercisesInWorkout)
                     .ThenInclude(eiw => eiw.Sets)
                 .OrderByDescending(completedWorkouts => completedWorkouts.Id)
@@ -114,12 +115,12 @@ namespace ftDB.Dao
             return workouts;
         }
 
-        public async Task<ModelPastWorkout> GetWorkoutAsync(int workoutId)
+        public async Task<ModelPastWorkout> GetWorkoutAsync(int workoutId, string uuid)
         {
             ModelPastWorkout workouts = await _context.CompletedWorkouts
                 .Include(cw => cw.ExercisesInWorkout)
                     .ThenInclude(eiw => eiw.Sets)
-                .Where(completedWorkouts => completedWorkouts.Id == workoutId)
+                .Where(completedWorkouts => completedWorkouts.Id == workoutId && completedWorkouts.Uuid == uuid)
                 .Select(completedWorkouts => new ModelPastWorkout(
                     completedWorkouts.Name,
                     completedWorkouts.Duration,
@@ -150,51 +151,13 @@ namespace ftDB.Dao
             return workouts;
         }
 
-        public async Task<ModelExerciseToUpdate> GetNewExerciseByIdAsync(int id)
-        {
-            ModelExerciseToUpdate? newExercise = await _context.ExercisesInWorkouts
-                                               .OrderByDescending(e => e.Id)
-                                               .Where(e => e.ExerciseId == id)
-                                               .Include(e => e.Exercise)
-                                               .Select(e => new ModelExerciseToUpdate(
-                                                   e.ExerciseId,
-                                                   e.Exercise.Name,
-                                                   e.Exercise.Equipment,
-                                                   e.Exercise.TargetMuscle,
-                                                   e.WeightUnit,
-                                                   "",
-                                                   new ModelSetToUpdate[]
-                                                   {
-                                                        new(-1, -1, 1, false),
-                                                   }
-                                               ))
-                                               .FirstOrDefaultAsync();
-
-            newExercise ??= await _context.Exercises
-                              .Where(e => e.Id == id)
-                              .Select(e => new ModelExerciseToUpdate(
-                                e.Id,
-                                e.Name,
-                                e.Equipment,
-                                e.TargetMuscle,
-                                "lbs",
-                                "",
-                                new ModelSetToUpdate[]
-                                {
-                                    new(-1, -1, 1, false),
-                                }
-                              )).FirstAsync();
-
-            return newExercise;
-        }
-
-        public async Task PostWorkoutTemplateAsync(RequestModelPostWorkoutTemplate workoutTemplate)
+        public async Task PostWorkoutTemplateAsync(RequestModelPostWorkoutTemplate workoutTemplate, string uuid)
         {
             using var transaction = _context.Database.BeginTransaction();
 
             try
             {
-                WorkoutTemplate templateToPost = new(workoutTemplate.Name);
+                WorkoutTemplate templateToPost = new(workoutTemplate.Name, uuid);
 
                 int postedTemplateId = await PostWorkoutTemplateToDBAsync(templateToPost);
 
@@ -232,12 +195,12 @@ namespace ftDB.Dao
             }
         }
 
-        public async Task<ModelGetWorkoutTemplate> GetWorkoutTemplateAsync(int workoutTemplateId)
+        public async Task<ModelGetWorkoutTemplate> GetWorkoutTemplateAsync(int workoutTemplateId, string uuid)
         {
             ModelGetWorkoutTemplate template = await _context.WorkoutTemplates
                 .Include(wt => wt.ExerciseTemplates)
                 .ThenInclude(et => et.SetTemplates)
-                .Where(workoutTemplate => workoutTemplate.Id == workoutTemplateId)
+                .Where(workoutTemplate => workoutTemplate.Id == workoutTemplateId && workoutTemplate.Uuid == uuid)
                 .Select(workoutTemplate => new ModelGetWorkoutTemplate(
                     workoutTemplate.Id,
                     workoutTemplate.Name,
@@ -265,9 +228,10 @@ namespace ftDB.Dao
             return template;
         }
 
-        public async Task<List<ModelGetWorkoutTemplate>> GetAllTemplatesAsync()
+        public async Task<List<ModelGetWorkoutTemplate>> GetAllTemplatesAsync(string uuid)
         {
             List<ModelGetWorkoutTemplate> templates = await _context.WorkoutTemplates
+                .Where(wt => wt.Uuid == uuid)
                 .Include(wt => wt.ExerciseTemplates)
                 .ThenInclude(et => et.SetTemplates)
                 .OrderByDescending(template => template.Id)
@@ -296,11 +260,11 @@ namespace ftDB.Dao
             return templates;
         }
 
-        public async Task<bool> DeleteWorkoutTemplateAsync(int workoutTemplateId)
+        public async Task<bool> DeleteWorkoutTemplateAsync(int workoutTemplateId, string uuid)
         {
             WorkoutTemplate? templateToDelete = await _context.WorkoutTemplates.FindAsync(workoutTemplateId);
 
-            if (templateToDelete != null)
+            if (templateToDelete != null && templateToDelete.Uuid == uuid)
             {
                 // Remove the entity from the DbSet
                 _context.WorkoutTemplates.Remove(templateToDelete);
@@ -314,11 +278,11 @@ namespace ftDB.Dao
             return false;
         }
 
-        public async Task<bool> DeleteWorkoutHistoryAsync(int workoutHistoryId)
+        public async Task<bool> DeleteWorkoutHistoryAsync(int workoutHistoryId, string uuid)
         {
             CompletedWorkout? historyToDelete = await _context.CompletedWorkouts.FindAsync(workoutHistoryId);
 
-            if (historyToDelete != null)
+            if (historyToDelete != null && historyToDelete.Uuid == uuid)
             {
                 // Remove the entity from the DbSet
                 _context.CompletedWorkouts.Remove(historyToDelete);
@@ -332,9 +296,9 @@ namespace ftDB.Dao
             return false;
         }
 
-        public async Task AddExerciseToDbAsync(RequestModelAddExercise exerciseToAdd)
+        public async Task AddExerciseToDbAsync(RequestModelAddExercise exerciseToAdd, string uuid)
         {
-            Exercise exercise = new(exerciseToAdd.Name, exerciseToAdd.Equipment, exerciseToAdd.TargetMuscle);
+            Exercise exercise = new(exerciseToAdd.Name, exerciseToAdd.Equipment, exerciseToAdd.TargetMuscle, uuid);
 
             _context.Exercises.Add(exercise);
 
@@ -343,7 +307,7 @@ namespace ftDB.Dao
             return;
         }
 
-        public async Task UpdateTemplateAsync(RequestModelUpdateTemplate modifiedTemplate)
+        public async Task UpdateTemplateAsync(RequestModelUpdateTemplate modifiedTemplate, string uuid)
         {
             using var transaction = _context.Database.BeginTransaction();
 
@@ -353,7 +317,7 @@ namespace ftDB.Dao
                 WorkoutTemplate? oldTemplate = await _context.WorkoutTemplates
                                 .Include(wt => wt.ExerciseTemplates)
                                 .ThenInclude(et => et.SetTemplates)
-                                .FirstOrDefaultAsync(workoutTemplate => workoutTemplate.Id == modifiedTemplate.Id);
+                                .FirstOrDefaultAsync(workoutTemplate => workoutTemplate.Id == modifiedTemplate.Id && workoutTemplate.Uuid == uuid);
 
                 if (oldTemplate == null)
                 {
@@ -457,7 +421,7 @@ namespace ftDB.Dao
             }
         }
 
-        public async Task UpdateHistoryAsync(RequestModelUpdateHistory completedWorkout)
+        public async Task UpdateHistoryAsync(RequestModelUpdateHistory completedWorkout, string uuid)
         {
             using var transaction = _context.Database.BeginTransaction();
 
@@ -467,7 +431,7 @@ namespace ftDB.Dao
                 CompletedWorkout? oldWorkout = await _context.CompletedWorkouts
                                 .Include(wt => wt.ExercisesInWorkout)
                                 .ThenInclude(et => et.Sets)
-                                .FirstOrDefaultAsync(workout => workout.Id == completedWorkout.Id);
+                                .FirstOrDefaultAsync(workout => workout.Id == completedWorkout.Id && workout.Uuid == uuid);
 
                 if (oldWorkout == null)
                 {
@@ -603,10 +567,10 @@ namespace ftDB.Dao
             await _context.SaveChangesAsync();
         }
 
-        private async Task<List<ModelExercise>> GetExerciseListWhenStringExistsAsync(string searchInput)
+        private async Task<List<ModelExercise>> GetExerciseListWhenStringExistsAsync(string searchInput, string uuid)
         {
             List<ModelExercise> exerciseList = await _context.Exercises
-                                                            .Where(exercise => exercise.SearchVector.Matches(searchInput))
+                                                            .Where(exercise => exercise.SearchVector.Matches(searchInput) && (exercise.Uuid == uuid || exercise.Uuid == null))
                                                             .OrderBy(exercise => exercise.Id)
                                                             .Select(exercise => new ModelExercise
                                                                     (
@@ -621,9 +585,10 @@ namespace ftDB.Dao
             return exerciseList;
         }
 
-        private async Task<List<ModelExercise>> GetExerciseListEmptyStringAsync()
+        private async Task<List<ModelExercise>> GetExerciseListEmptyStringAsync(string uuid)
         {
             List<ModelExercise> exerciseList = await _context.Exercises
+                                                            .Where(exercise => exercise.Uuid == uuid || exercise.Uuid == null)
                                                             .OrderBy(exercise => exercise.Id)
                                                             .Select(exercise => new ModelExercise
                                                                     (
